+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Multiple bucket ID support to Share.
+
+### Fixed
+
+- Value of `TransferMethod.SOURCE_URL`.
+
 ## [3.12.0] - 2024-07-23
 
 ### Added